requirements.txt
asgiref==3.2.5
Django==3.0.4
djangorestframework==3.11.0
django-cors-headers==3.2.1
Pillow==3.4.0
psycopg2==2.8.4
python-dotenv==0.12.0
pytz==2019.3
requests==2.24.0
rcssmin==1.0.6
rjsmin==1.1.0
six==1.14.0
sqlparse==0.3.1
openpay==0.6
